requirements.txt
flask==3.1.0
groq==0.25.0
python-dotenv==1.1.0
gunicorn==21.2.0
